refactor(site/demo): extract REPL shipment demo to a real .vue + ?raw import
The shipment demo was previously a 1500-line template literal inside
DemoReplEditor.client.vue. Vue-tsc and eslint both saw it as opaque
text, so type-level regressions slipped past every static check —
that's how the form.undo() / form.history.undo() rename went out as
a silently-broken demo.

Move the demo source into apps/site/repl-demos/shipment-demo.vue —
a real Vue SFC outside the components/ scanning path — and import its
text via Vite's `?raw` query. The same source is now type-checked at
build time AND served to users; there's no extraction step and no
drift between what we review and what we ship.

Two type issues that the extracted file surfaced, both fixed here:
  - The cargo-items v-for iterated `form.values.cargo.items` for the
    index, then read `form.fields.cargo.items[idx]` inside. TS flagged
    the indexed-access as possibly-undefined. Switch to iterate
    `form.fields.cargo.items` directly and reference each `item`
    inline — reads cleaner anyway.
  - `removeLineItem` takes `number`; v-for over the FieldStateMap
    yields a template-literal `${number}` key. Coerce at the call
    site with `Number(idx)`.
